+++ pkg/integrations/prometheus/utils/operatorutils.go
@@ -260,17 +260,3 @@ func ExtractMatchingLabelSet(groupIDStr string, matchingLabels []string, include
 
 	return matchingLabelValStr
 }
-
-func CalculateStandardVariance(values []float64) float64 {
-	sum := 0.0
-	length := float64(len(values))
-	for _, val := range values {
-		sum += val
-	}
-	avg := sum / length
-	sumValSquare := 0.0
-	for _, val := range values {
-		sumValSquare += (val - avg) * (val - avg)
-	}
-	return sumValSquare / length
-}
